refactor: Member 관련 DTO와 Converter 구조 정리 (#14)
