+++ src/main/java/umc/spring/converter/MemberConverter.java
@@ -1,24 +1,32 @@
 package umc.spring.converter;
 
 import lombok.Getter;
+import org.springframework.data.domain.Page;
 import umc.spring.domain.Member;
+import umc.spring.domain.Review;
 import umc.spring.domain.enums.Gender;
+import umc.spring.domain.enums.MissionStatus;
+import umc.spring.domain.mapping.MemberMission;
 import umc.spring.web.dto.MemberRequestDto;
 import umc.spring.web.dto.MemberResponseDto;
+import umc.spring.web.dto.MissionResponseDto;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
 
 public class MemberConverter {
 
-    public static MemberResponseDto.JoinResultDTO joinResultDTO(Member member) {
+
+    public static MemberResponseDto.JoinResultDTO joinResultDTO (Member member) {
         return MemberResponseDto.JoinResultDTO.builder()
                 .memberId(member.getId())
                 .createdAt(LocalDateTime.now())
                 .build();
     }
 
-    public static Member toMember(MemberRequestDto.JoinDto request) {
+    public static Member toMember (MemberRequestDto.JoinDto request) {
 
         Gender gender = null;
 
@@ -42,5 +50,66 @@ public class MemberConverter {
                 .memberPreferList(new ArrayList<>())
                 .build();
     }
+
+
+
+    public static MemberResponseDto.reviewReviewDto reviewReviewDto (Review review) {
+
+        return MemberResponseDto.reviewReviewDto.builder()
+                .storeName(review.getStore().getName())
+                .ownerNickname(review.getMember().getName())
+                .body(review.getBody())
+                .score(review.getScore())
+                .createdAt(review.getCreatedAt().toLocalDate())
+                .build();
+
+    }
+
+    public static MemberResponseDto.reviewReviewListDto reviewReviewListDto (Page<Review> pageReviewList) {
+        List<MemberResponseDto.reviewReviewDto> reviewReviewDtoList
+                = pageReviewList.stream().map(MemberConverter::reviewReviewDto).collect(Collectors.toList());
+
+        return MemberResponseDto.reviewReviewListDto.builder()
+                .reviewReviewDtoList(reviewReviewDtoList)
+                .listSize(pageReviewList.getSize())
+                .totalPage(pageReviewList.getTotalPages())
+                .totalElements(pageReviewList.getTotalElements())
+                .isFirst(pageReviewList.isFirst())
+                .isLast(pageReviewList.isLast())
+                .build();
+
+    }
+
+
+    public static MemberResponseDto.memberMissionPreviewDto memberMissionPreviewDto (MemberMission memberMission) {
+
+        return MemberResponseDto.memberMissionPreviewDto.builder()
+                .storeName(memberMission.getMission().getStore().getName())
+                .missionSpec(memberMission.getMission().getMissionSpec())
+                .reward(memberMission.getMission().getReward())
+                .missionStatus(memberMission.getMissionStatus())
+                .build();
+
+    }
+
+    public static MemberResponseDto.memberMissionListPreviewDto memberMissionListInChallengingPreviewDto (Page<MemberMission> pageMemberMissionList) {
+
+        List<MemberResponseDto.memberMissionPreviewDto> memberMissionPreviewDtoList
+                = pageMemberMissionList.stream()
+                .filter(memberMission -> MissionStatus.CHALLENGING.equals(memberMission.getMissionStatus()))
+                .map(MemberConverter::memberMissionPreviewDto)
+                .collect(Collectors.toList());
+
+        return MemberResponseDto.memberMissionListPreviewDto.builder()
+                .memberMissionList(memberMissionPreviewDtoList)
+                .listSize(pageMemberMissionList.getSize())
+                .totalElements(pageMemberMissionList.getTotalElements())
+                .isFirst(pageMemberMissionList.isFirst())
+                .isLast(pageMemberMissionList.isLast())
+                .totalPage(pageMemberMissionList.getTotalPages())
+                .build();
+
+    }
+
 
 }
