+++ src/main/java/umc/spring/converter/MemberMissionConverter.java
@@ -12,42 +12,7 @@ import java.util.stream.Collectors;
 
 public class MemberMissionConverter {
 
-
-    public static MemberResponseDto.memberMissionPreviewDto memberMissionPreviewDto (MemberMission memberMission) {
-
-        return MemberResponseDto.memberMissionPreviewDto.builder()
-                .storeName(memberMission.getMission().getStore().getName())
-                .missionSpec(memberMission.getMission().getMissionSpec())
-                .reward(memberMission.getMission().getReward())
-                .missionStatus(memberMission.getMissionStatus())
-                .build();
-
-    }
-
-    public static MemberResponseDto.memberMissionListPreviewDto memberMissionListInChallengingPreviewDto (Page<MemberMission> pageMemberMissionList) {
-
-        List<MemberResponseDto.memberMissionPreviewDto> memberMissionPreviewDtoList
-                = pageMemberMissionList.stream()
-                .filter(memberMission -> MissionStatus.CHALLENGING.equals(memberMission.getMissionStatus()))
-                .map(MemberMissionConverter::memberMissionPreviewDto)
-                .collect(Collectors.toList());
-
-        return MemberResponseDto.memberMissionListPreviewDto.builder()
-                .memberMissionList(memberMissionPreviewDtoList)
-                .listSize(pageMemberMissionList.getSize())
-                .totalElements(pageMemberMissionList.getTotalElements())
-                .isFirst(pageMemberMissionList.isFirst())
-                .isLast(pageMemberMissionList.isLast())
-                .totalPage(pageMemberMissionList.getTotalPages())
-                .build();
-
-    }
-
-
-
-
     public static MemberMissionDto.resultAddMemberMission resultAddMemberMission(MemberMission memberMission) {
-
 
         return MemberMissionDto.resultAddMemberMission.builder()
                 .memberMissionId(memberMission.getId())
@@ -57,7 +22,7 @@ public class MemberMissionConverter {
 
     }
 
-    public static MemberMission toMemberMission (MemberMissionDto.addMemberMissionDto addMemberMissionDto) {
+    public static MemberMission toMemberMission () {
         return MemberMission.builder()
                 .missionStatus(MissionStatus.CHALLENGING)
                 .build();
