+++ src/main/java/umc/spring/converter/MissionConverter.java
@@ -11,32 +11,6 @@ import java.util.stream.Collectors;
 
 public class MissionConverter {
 
-    public static MissionResponseDto.reviewReviewDto reviewReviewDto (Review review) {
-
-        return MissionResponseDto.reviewReviewDto.builder()
-                .storeName(review.getStore().getName())
-                .ownerNickname(review.getMember().getName())
-                .body(review.getBody())
-                .score(review.getScore())
-                .createdAt(review.getCreatedAt().toLocalDate())
-                .build();
-
-    }
-
-    public static MissionResponseDto.reviewReviewListDto reviewReviewListDto(Page<Review> pageReviewList) {
-        List<MissionResponseDto.reviewReviewDto> reviewReviewDtoList
-                = pageReviewList.stream().map(MissionConverter::reviewReviewDto).collect(Collectors.toList());
-
-        return MissionResponseDto.reviewReviewListDto.builder()
-                .reviewReviewDtoList(reviewReviewDtoList)
-                .listSize(pageReviewList.getSize())
-                .totalPage(pageReviewList.getTotalPages())
-                .totalElements(pageReviewList.getTotalElements())
-                .isFirst(pageReviewList.isFirst())
-                .isLast(pageReviewList.isLast())
-                .build();
-
-    }
 
     public static MissionResponseDto.AddMissionResultDto addMissionResultDto(Mission mission) {
