--- src/main/java/umc/spring/converter/RegionStoreListConverter.java
@@ -1,4 +1,0 @@
-package umc.spring.converter;
-
-public class RegionStoreListConverter {
-}
