+++ src/main/java/umc/spring/validation/annotation/CheckPage.java
@@ -12,7 +12,6 @@ import java.lang.annotation.*;
 @Target( { ElementType.METHOD, ElementType.FIELD, ElementType.PARAMETER })
 @Retention(RetentionPolicy.RUNTIME)
 public @interface CheckPage {
-
     String message() default "해당 페이지는 존재하지 않습니다.";
     Class<?>[] groups() default {};
     Class<? extends Payload>[] payload() default {};
