+++ src/main/java/umc/spring/validation/annotation/ExistStore.java
@@ -11,7 +11,6 @@ import java.lang.annotation.*;
 @Target( { ElementType.METHOD, ElementType.FIELD, ElementType.PARAMETER })
 @Retention(RetentionPolicy.RUNTIME)
 public @interface ExistStore {
-
     String message() default "해당 가게가 존재하지 않습니다.";
     Class<?>[] groups() default {};
     Class<? extends Payload>[] payload() default {};
