+++ src/main/java/umc/spring/web/controller/MemberController.java
@@ -43,7 +43,7 @@ public class MemberController {
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "AUTH004", description = "acess 토큰 만료",content = @Content(schema = @Schema(implementation = ApiResponse.class))),
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "AUTH006", description = "acess 토큰 모양이 이상함",content = @Content(schema = @Schema(implementation = ApiResponse.class))),
     })
-    public ApiResponse<MemberResponseDto.JoinResultDTO> join (@ExistCategories @RequestBody MemberRequestDto.JoinDto request) {
+    public ApiResponse<MemberResponseDto.JoinResultDTO> join (@RequestBody MemberRequestDto.JoinDto request) {
         Member member = memberCommandService.joinMember(request);
 
         return ApiResponse.onSuccess(MemberConverter.joinResultDTO(member));
