+++ src/main/java/umc/spring/web/dto/StoreResponseDto.java
@@ -79,4 +79,6 @@ public class StoreResponseDto {
         LocalDate createdAt;
 
     }
+
+
 }
